case_study_generator/requirements.txt
# Core dependencies
google-generativeai>=0.3.0
openai>=1.0.0
notion-client>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Development dependencies
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0

GOOGLE_API_KEY=your_gemini_api_key
OPENAI_API_KEY=your_openai_api_key
NOTION_API_KEY=your_notion_api_key
NOTION_DATABASE_ID=your_notion_database_id 